+++ src/bin/commands/mod.rs
@@ -1,3 +1,7 @@
+mod clean;
+mod containers;
 mod images;
 
+pub use self::clean::*;
+pub use self::containers::*;
 pub use self::images::*;