+++ src/bin/cross-util.rs
@@ -1,6 +1,7 @@
 #![deny(missing_debug_implementations, rust_2018_idioms)]
 
 use clap::{CommandFactory, Parser, Subcommand};
+use cross::docker;
 
 mod commands;
 
@@ -23,9 +24,17 @@ struct CliHidden {
 
 #[derive(Subcommand, Debug)]
 enum Commands {
-    /// List cross images in local storage.
+    /// Work with cross images in local storage.
     #[clap(subcommand)]
     Images(commands::Images),
+    /// Work with cross volumes in local storage.
+    #[clap(subcommand)]
+    Volumes(commands::Volumes),
+    /// Work with cross containers in local storage.
+    #[clap(subcommand)]
+    Containers(commands::Containers),
+    /// Clean all cross data in local storage.
+    Clean(commands::Clean),
 }
 
 fn is_toolchain(toolchain: &str) -> cross::Result<String> {
@@ -37,16 +46,13 @@ fn is_toolchain(toolchain: &str) -> cross::Result<String> {
     }
 }
 
-fn get_container_engine(
-    engine: Option<&str>,
-    verbose: bool,
-) -> cross::Result<cross::docker::Engine> {
+fn get_container_engine(engine: Option<&str>, verbose: bool) -> cross::Result<docker::Engine> {
     let engine = if let Some(ce) = engine {
         which::which(ce)?
     } else {
-        cross::docker::get_container_engine()?
+        docker::get_container_engine()?
     };
-    cross::docker::Engine::from_path(engine, verbose)
+    docker::Engine::from_path(engine, None, verbose)
 }
 
 pub fn main() -> cross::Result<()> {
@@ -55,6 +61,18 @@ pub fn main() -> cross::Result<()> {
     match cli.command {
         Commands::Images(args) => {
             let engine = get_container_engine(args.engine(), args.verbose())?;
+            args.run(engine)?;
+        }
+        Commands::Volumes(args) => {
+            let engine = get_container_engine(args.engine(), args.verbose())?;
+            args.run(engine, cli.toolchain.as_deref())?;
+        }
+        Commands::Containers(args) => {
+            let engine = get_container_engine(args.engine(), args.verbose())?;
+            args.run(engine)?;
+        }
+        Commands::Clean(args) => {
+            let engine = get_container_engine(args.engine.as_deref(), args.verbose)?;
             args.run(engine)?;
         }
     }