+++ src/docker/engine.rs
@@ -2,6 +2,7 @@ use std::env;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
+use crate::config::bool_from_envvar;
 use crate::errors::*;
 use crate::extensions::CommandExt;
 
@@ -20,19 +21,35 @@ pub enum EngineType {
 pub struct Engine {
     pub kind: EngineType,
     pub path: PathBuf,
+    pub is_remote: bool,
 }
 
 impl Engine {
-    pub fn new(verbose: bool) -> Result<Engine> {
+    pub fn new(is_remote: Option<bool>, verbose: bool) -> Result<Engine> {
         let path = get_container_engine()
             .map_err(|_| eyre::eyre!("no container engine found"))
             .with_suggestion(|| "is docker or podman installed?")?;
-        Self::from_path(path, verbose)
+        Self::from_path(path, is_remote, verbose)
     }
 
-    pub fn from_path(path: PathBuf, verbose: bool) -> Result<Engine> {
+    pub fn from_path(path: PathBuf, is_remote: Option<bool>, verbose: bool) -> Result<Engine> {
         let kind = get_engine_type(&path, verbose)?;
-        Ok(Engine { path, kind })
+        let is_remote = is_remote.unwrap_or_else(Self::is_remote);
+        Ok(Engine {
+            path,
+            kind,
+            is_remote,
+        })
+    }
+
+    pub fn needs_remote(&self) -> bool {
+        self.is_remote && self.kind == EngineType::Podman
+    }
+
+    pub fn is_remote() -> bool {
+        env::var("CROSS_REMOTE")
+            .map(|s| bool_from_envvar(&s))
+            .unwrap_or_default()
     }
 }